roles

+++ e-commerce/ViewModels/CartVM.cs
@@ -1,11 +1,13 @@
 ﻿using e_commerce.Models;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace e_commerce.ViewModels
 {
     public class CartVM
     {
         public Cart Cart { get; set; }
-        public double CartTotal { get; set; }
+        public decimal CartTotal { get; set; }
 
     }
 }
